move large example into batch exchange
